x/canvas/types/message_create_canvas_test.go
package types

import (
	"testing"

	"canvas/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateCanvas_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateCanvas
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateCanvas{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateCanvas{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
